requirements.txt
python-telegram-bot
httpx
feedparser
beautifulsoup4
python-dotenv